% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_script.R
\name{check_bids_path}
\alias{check_bids_path}
\title{Tests if the BIDS folder contains a "participants.tsv" file}
\usage{
check_bids_path(path = bids_path)
}
\description{
Tests if the BIDS folder contains a "participants.tsv" file
}
