% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_script.R
\name{create_bids_df}
\alias{create_bids_df}
\title{Creates a BIDS dataframe with all relevant information}
\usage{
create_bids_df(path = bids_path, output = output_path)
}
\description{
Creates a BIDS dataframe with all relevant information
}
