% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_script.R
\name{create_participants_tsv}
\alias{create_participants_tsv}
\title{Creates an adapted participants.tsv in the output path}
\usage{
create_participants_tsv(
  path = output_path,
  df = df_bids_data_filter,
  session = filter_sessions,
  sequences = filter_sequences,
  cohort = filter_cohort
)
}
\arguments{
\item{df}{}

\item{session}{}

\item{sequences}{}

\item{cohort}{}

\item{output_path}{}
}
\description{
Creates an adapted participants.tsv in the output path
}
