% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_script.R
\name{filter_dialog}
\alias{filter_dialog}
\title{Starts the "filter dialog" for sequence, session and regular expressions in the BIDS dataset}
\usage{
filter_dialog()
}
\description{
Starts the "filter dialog" for sequence, session and regular expressions in the BIDS dataset
}
