% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_script.R
\name{test_filter}
\alias{test_filter}
\title{Checks the filter criterias}
\usage{
test_filter(
  df,
  session = filter_sessions,
  sequences = filter_sequences,
  cohort = filter_cohort
)
}
\arguments{
\item{df}{}

\item{session}{}

\item{sequences}{}

\item{cohort}{}
}
\description{
Checks the filter criterias
}
